Fix notice components error

+++ components/notice/noticeService.js
@@ -1,4 +1,5 @@
 const noticeProvider = require('./noticeProvider');
+const userProvider = require('../user/userProvider');
 const noticeDao = require('./noticeDao');
 const { response, errResponse } = require('../../config/response');
 const resStatus = require('../../config/resStatus');
@@ -6,16 +7,19 @@ const withConnection = require('../../config/connection')
 
 exports.posting = withConnection(async (connection, userEmail, title, contents) => {
     const userInfo = await userProvider.adminbyEmail(userEmail);
-    if (userInfo.length < 1) return errResponse(resStatus.USER_USEREMAIL_NOT_EXIST);
+    if (!userInfo) return errResponse(resStatus.USER_ADMINEMAIL_NOT_EXIST);
 
     const newNoticeInfo = [title, contents]; //Title, Contents
 
-    const newNoticeID = await noticeDao.insertNotice(connection, newNoticeInfo);
+    const newNoticeHeader = await noticeDao.insertNotice(connection, newNoticeInfo);
 
-    return response(resStatus.NOTICE_POSTING_SUCCESS, { "noticeID": newNoticeID.toString() }); 
+    return response(resStatus.NOTICE_POSTING_SUCCESS, { "noticeID": newNoticeHeader.insertId }); 
 });
 
-exports.editNotice = withConnection(async (connection, noticeId, title, contents) => {
+exports.editNotice = withConnection(async (connection, userEmail, noticeId, title, contents) => {
+    const userInfo = await userProvider.adminbyEmail(userEmail);
+    if (!userInfo) return errResponse(resStatus.USER_ADMINEMAIL_NOT_EXIST);
+
     const selectedNotice = await noticeProvider.itembyNoticeId(noticeId);
     if (selectedNotice.length < 1) return errResponse(resStatus.NOTICE_NOT_EXIST);
 
@@ -23,16 +27,18 @@ exports.editNotice = withConnection(async (connection, noticeId, title, contents
 
     const editedNoticeID = await noticeDao.updateNotice(connection, editedNoticeInfo);
 
-    return response(resStatus.NOTICE_EDIT_SUCCESS, { "noticeID": editedNoticeID.toString() });
+    return response(resStatus.NOTICE_EDIT_SUCCESS, { "noticeID": noticeId });
 });
 
-exports.deleteNotice = withConnection(async (connection, noticeId) => {
+exports.deleteNotice = withConnection(async (connection, userEmail, noticeId) => {
+    const userInfo = await userProvider.adminbyEmail(userEmail);
+    if (!userInfo) return errResponse(resStatus.USER_ADMINEMAIL_NOT_EXIST);
 
     const selectedNotice = await noticeProvider.itembyNoticeId(noticeId);
     if (selectedNotice.length < 1) return errResponse(resStatus.NOTICE_NOT_EXIST);
 
     const deletedNoticeID = await noticeDao.deleteNotice(connection, noticeId);
 
-    return response(resStatus.NOTICE_DELETE_SUCCESS, { "noticeID": deletedNoticeID.toString() });
+    return response(resStatus.NOTICE_DELETE_SUCCESS, { "noticeID": noticeId });
 
 });
